Bug 512525 - Editor actions in menus and Generic Editor
* Add menus so commands are accessible via shortcut and quick-assist
* Add images to commands so they look nice in menus
* Test enablement of wordwrap, whitesplaces, block selection in Generic
Editor...

Bug: 512525
Bug: 512571

+++ org.eclipse.ui.genericeditor.tests/src/org/eclipse/ui/genericeditor/tests/AbstratGenericEditorTest.java
@@ -12,6 +12,9 @@ package org.eclipse.ui.genericeditor.tests;
 
 import java.io.ByteArrayInputStream;
 
+import org.junit.After;
+import org.junit.Before;
+
 import org.eclipse.swt.widgets.Display;
 
 import org.eclipse.core.resources.IFile;
@@ -19,35 +22,52 @@ import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.ResourcesPlugin;
 
 import org.eclipse.ui.PlatformUI;
+import org.eclipse.ui.internal.genericeditor.ExtensionBasedTextEditor;
 import org.eclipse.ui.intro.IIntroPart;
+import org.eclipse.ui.part.FileEditorInput;
 
-public class GenericEditorTestUtils {
+/**
+ * Closes intro, create {@link #project}, create {@link #file} and open {@link #editor}; and clean up.
+ * Also contains additional utility methods
+ * @since 1.0 
+ */
+public class AbstratGenericEditorTest {
 	
-	private static IProject project;
-	private static IFile file;
+	protected IProject project;
+	protected IFile file;
+	protected ExtensionBasedTextEditor editor;
 
-	 public static void setUpBeforeClass() throws Exception {
-		project = ResourcesPlugin.getWorkspace().getRoot().getProject("genericEditorTest");
+	/**
+	 * Closes intro, create {@link #project}, create {@link #file} and open {@link #editor}
+	 * @throws Exception ex
+	 */
+	@Before
+	public void setUp() throws Exception {
+		closeIntro();
+		project = ResourcesPlugin.getWorkspace().getRoot().getProject(getClass().getName() + System.currentTimeMillis());
 		project.create(null);
 		project.open(null);
 		file = project.getFile("foo.txt");
 		file.create(new ByteArrayInputStream("bar 'bar'".getBytes()), true, null);
+		editor = (ExtensionBasedTextEditor) PlatformUI.getWorkbench().getActiveWorkbenchWindow()
+				.getActivePage().openEditor(new FileEditorInput(this.file), "org.eclipse.ui.genericeditor.GenericEditor");
 	 }
 
-	public static void tearDownAfterClass() throws Exception {
-		file.delete(true, null);
-		project.delete(true, null);
+	@After
+	public void tearDown() throws Exception {
+		if (file != null) {
+			file.delete(true, null);
+		}
+		if (project != null) {
+			project.delete(true, null);
+		}
 	}
 
-	public static void closeIntro() {
+	private static void closeIntro() {
 		IIntroPart intro = PlatformUI.getWorkbench().getIntroManager().getIntro();
 		if (intro != null) {
 			PlatformUI.getWorkbench().getIntroManager().closeIntro(intro);
 		}
-	}
-	
-	public static IFile getFile(){
-		return file;
 	}
 	
 	public static void waitAndDispatch(long milliseconds) {
